ui수정 및 삭제수정

+++ src/main/java/my/example/spring/diary/dao/BoardDaoSQL.java
@@ -3,7 +3,7 @@ package my.example.spring.diary.dao;
 public class BoardDaoSQL {
 
     public static final String SELECT_BOARDS =
-            "SELECT id, nickname,content,regdate FROM board limit :start,:limit";
+            "SELECT id, nickname,content,regdate FROM board WHERE user_id=:userid limit :start,:limit";
 
     public static final String SELECTBOARDS =
             "SELECT id, nickname,content,regdate FROM board ";
@@ -12,10 +12,10 @@ public class BoardDaoSQL {
             "SELECT id,nickname,content,regdate FROM board where id=:id";
 
     public static final String UPDATE_BOARD =
-            "UPDATE board SET title=:title,content=:content WHERE id=:id";
+            "UPDATE board SET content=:content WHERE id=:id";
 
     public static final String DELETE_BOARD =
-            "DELETE FROM board WHERE id=:id";
+            "DELETE FROM board WHERE id=:id and user_id= :userid";
 
     public static final String INSERT =
            "INSERT INTO board(nickname, content) VALUES(?, ?)";
